requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==23.5
matplotlib==3.8.2
pandas==2.1.4
mercadopago==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0